+++ pytket/pytket/qasm/qasm.py
@@ -335,6 +335,7 @@ class CircuitTransformer(Transformer):
         except StopIteration:
             args = next_tree
             pars = []
+        args = list(args)
         params = [f"({par})/pi" for par in pars]
         if opstr in self.gate_dict:
             gdef = self.gate_dict[opstr]
@@ -357,11 +358,12 @@ class CircuitTransformer(Transformer):
             op = {"type": optype}
             if params:
                 op["params"] = params
+            # Operations needing special handling:
             if optype.startswith("Cn"):
-                # n-controlled rotation are only gates supported
-                # via this function without fixed signature
-                args = list(args)
+                # n-controlled rotations have variable signature
                 op["n_qb"] = len(args)
+            elif optype == "Barrier":
+                op["signature"] = ["Q"] * len(args)
 
         for arg in zip(*self.unroll_all_args(args)):
             yield {"args": list(arg), "op": op}